requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.8.0
scikit-learn>=1.0.0
cvxpy>=1.2.0
tqdm>=4.62.0

# Data and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
wrds>=3.1.0

# AI/LangChain dependencies (updated to latest modular structure)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
huggingface_hub>=0.19.0
hf-xet>=0.1.0

# Web application dependencies (planned)
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Development dependencies (optional)
pytest>=6.0.0
jupyter>=1.0.0
notebook>=6.0.0